[refactor] 디렉토리 구조 변경 - domain 디렉토리 하위에 WebSearch, SearchInfo, Chat 관리 - Resolves: #29
