+++ news-api/src/main/kotlin/com/search/config/web/http/WebConfig.kt
@@ -1,0 +1,18 @@
+package com.search.config.web.http
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
+
+@Configuration
+class WebConfig: WebMvcConfigurer {
+
+    override fun addCorsMappings(registry: CorsRegistry) {
+        registry.addMapping("/**")
+                .allowedOriginPatterns("http://localhost:3000")
+                .allowedMethods("GET", "POST", "PUT", "DELETE")
+                .allowedHeaders("Content-Type")
+                .allowCredentials(true)
+                .maxAge(3600);
+    }
+}
