+++ news-api/src/main/kotlin/com/search/config/web/websocket/WebSocketConfig.kt
@@ -1,0 +1,17 @@
+package com.search.config.web.websocket
+
+import com.search.config.web.websocket.handler.MyWebSocketHandler
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.socket.config.annotation.EnableWebSocket
+import org.springframework.web.socket.config.annotation.WebSocketConfigurer
+import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry
+
+@Configuration
+@EnableWebSocket
+class WebSocketConfig(
+        private val myWebSocketHandler: MyWebSocketHandler
+): WebSocketConfigurer{
+    override fun registerWebSocketHandlers(registry: WebSocketHandlerRegistry) {
+        registry.addHandler(myWebSocketHandler, "/connect")
+    }
+}
