+++ news-api/src/main/kotlin/com/search/config/web/websocket/handler/MyWebSocketHandler.kt
@@ -1,0 +1,14 @@
+package com.search.config.web.websocket.handler
+
+import org.springframework.stereotype.Component
+import org.springframework.web.socket.handler.TextWebSocketHandler
+
+/***
+ * connect 웹소켓 연결요청이 들어올때 이를 처리할 클래스
+ */
+
+@Component
+class MyWebSocketHandler: TextWebSocketHandler() {
+
+
+}
