+++ news-api/src/main/kotlin/com/search/domain/searchinfo/infrastructure/result/SearchInfoQueryResult.kt
@@ -1,0 +1,7 @@
+package com.search.domain.searchinfo.infrastructure.result
+
+class SearchInfoQueryResult(
+        var id: Int,
+        var query: String
+) {
+}
