+++ news-api/src/main/kotlin/com/search/domain/searchinfo/infrastructure/result/TopQueryResult.kt
@@ -1,0 +1,6 @@
+package com.search.domain.searchinfo.infrastructure.result
+
+data class TopQueryResult(
+        var query: String,
+        var count: Int
+)
