+++ news-api/src/main/kotlin/com/search/domain/websearch/controller/response/PageSearchResponse.kt
@@ -1,0 +1,7 @@
+package com.search.domain.websearch.controller.response
+
+data class PageSearchResponse<T>(val page: Int,
+                                 val size: Int,
+                                 val totalElements: Int,
+                                 val contents: List<T>
+)
