+++ news-api/src/main/kotlin/com/search/domain/websearch/controller/response/SearchResponse.kt
@@ -1,0 +1,7 @@
+package com.search.domain.websearch.controller.response
+
+data class SearchResponse(
+        var title: String,
+        var link: String,
+        var description: String
+)
