+++ news-api/src/main/kotlin/com/search/domain/websearch/controller/response/TopRankResponse.kt
@@ -1,0 +1,6 @@
+package com.search.domain.websearch.controller.response
+
+data class TopRankResponse(
+        var query: String,
+        var count: Int
+)
