+++ news-api/src/main/kotlin/com/search/domain/websearch/event/EventRequest.kt
@@ -1,0 +1,8 @@
+package com.search.domain.websearch.event
+
+import java.time.LocalDateTime
+
+data class EventRequest(
+        var query: String,
+        var timestamp: LocalDateTime
+)
