+++ news-api/src/main/kotlin/com/search/domain/websearch/infrastructure/result/WebSearchPageResult.kt
@@ -1,0 +1,7 @@
+package com.search.domain.websearch.infrastructure.result
+
+data class WebSearchPageResult<T>(val page: Int,
+                                  val size: Int,
+                                  val totalElements: Int,
+                                  val contents: List<T>
+)
