+++ news-api/src/main/kotlin/com/search/domain/websearch/infrastructure/result/WebSearchResult.kt
@@ -1,0 +1,7 @@
+package com.search.domain.websearch.infrastructure.result
+
+data class WebSearchResult(
+        val title: String,
+        val link: String,
+        val description: String
+)
