--- news-api/src/main/kotlin/com/search/searchinfo/infrastructure/result/SearchInfoQueryResult.kt
@@ -1,7 +1,0 @@
-package com.search.searchinfo.infrastructure.result
-
-class SearchInfoQueryResult(
-        var id: Int,
-        var query: String
-) {
-}
