--- news-api/src/main/kotlin/com/search/searchinfo/infrastructure/result/TopQueryResult.kt
@@ -1,6 +1,0 @@
-package com.search.searchinfo.infrastructure.result
-
-data class TopQueryResult(
-        var query: String,
-        var count: Int
-)
