--- news-api/src/main/kotlin/com/search/websearch/controller/response/PageSearchResponse.kt
@@ -1,7 +1,0 @@
-package com.search.websearch.controller.response
-
-data class PageSearchResponse<T>(val page: Int,
-                                 val size: Int,
-                                 val totalElements: Int,
-                                 val contents: List<T>
-)
