--- news-api/src/main/kotlin/com/search/websearch/controller/response/SearchResponse.kt
@@ -1,7 +1,0 @@
-package com.search.websearch.controller.response
-
-data class SearchResponse(
-        var title: String,
-        var link: String,
-        var description: String
-)
