--- news-api/src/main/kotlin/com/search/websearch/controller/response/TopRankResponse.kt
@@ -1,6 +1,0 @@
-package com.search.websearch.controller.response
-
-data class TopRankResponse(
-        var query: String,
-        var count: Int
-)
