--- news-api/src/main/kotlin/com/search/websearch/event/EventRequest.kt
@@ -1,8 +1,0 @@
-package com.search.websearch.event
-
-import java.time.LocalDateTime
-
-data class EventRequest(
-        var query: String,
-        var timestamp: LocalDateTime
-)
