--- news-api/src/main/kotlin/com/search/websearch/infrastructure/result/WebSearchPageResult.kt
@@ -1,7 +1,0 @@
-package com.search.websearch.infrastructure.result
-
-data class WebSearchPageResult<T>(val page: Int,
-                                  val size: Int,
-                                  val totalElements: Int,
-                                  val contents: List<T>
-)
