--- news-api/src/main/kotlin/com/search/websearch/infrastructure/result/WebSearchResult.kt
@@ -1,7 +1,0 @@
-package com.search.websearch.infrastructure.result
-
-data class WebSearchResult(
-        val title: String,
-        val link: String,
-        val description: String
-)
