--- news-api/src/main/kotlin/com/search/websearch/service/port/WebRepository.kt
@@ -1,8 +1,0 @@
-package com.search.websearch.service.port
-
-import com.search.websearch.infrastructure.result.WebSearchPageResult
-import com.search.websearch.infrastructure.result.WebSearchResult
-
-interface WebRepository {
-    fun search(query: String, page: Int, size: Int): WebSearchPageResult<WebSearchResult>
-}
